refactor: move ImplicitLinkedHash* to internals
