+++ clients/src/main/java/org/apache/kafka/common/utils/internals/ImplicitLinkedHashCollection.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.kafka.common.utils;
+package org.apache.kafka.common.utils.internals;
 
 import java.util.AbstractCollection;
 import java.util.AbstractSequentialList;
