+++ clients/src/test/java/org/apache/kafka/common/utils/internals/ImplicitLinkedHashMultiCollectionTest.java
@@ -14,9 +14,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kafka.common.utils;
+package org.apache.kafka.common.utils.internals;
 
-import org.apache.kafka.common.utils.ImplicitLinkedHashCollectionTest.TestElement;
+import org.apache.kafka.common.utils.internals.ImplicitLinkedHashCollectionTest.TestElement;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.Timeout;
