+++ server/src/main/java/org/apache/kafka/server/FetchSessionCacheShard.java
@@ -17,7 +17,7 @@
 package org.apache.kafka.server;
 
 import org.apache.kafka.common.requests.FetchMetadata;
-import org.apache.kafka.common.utils.ImplicitLinkedHashCollection;
+import org.apache.kafka.common.utils.internals.ImplicitLinkedHashCollection;
 import org.apache.kafka.common.utils.LogContext;
 import org.apache.kafka.server.FetchSession.EvictableKey;
 import org.apache.kafka.server.FetchSession.LastUsedKey;
