+++ server/src/main/java/org/apache/kafka/server/share/session/ShareSession.java
@@ -18,7 +18,7 @@
 package org.apache.kafka.server.share.session;
 
 import org.apache.kafka.common.TopicIdPartition;
-import org.apache.kafka.common.utils.ImplicitLinkedHashCollection;
+import org.apache.kafka.common.utils.internals.ImplicitLinkedHashCollection;
 import org.apache.kafka.server.share.CachedSharePartition;
 
 import java.util.ArrayList;
